Refactor FilesView into modular components
